Find timestamp of last executed system upgrade
Every time `pacman -Su` is executed, the line
"starting full system upgrade" is logged in /var/log/pacman.log,
regardless if the user proceeds or not at prompt. When the system
upgrade is started by the user, the line "transaction started"
immediately follows in /var/log/pacman.log, making it a better
indication that the user actually started the update.